backend/requirements.txt
fastapi~=0.81.0
pydantic~=1.10.1
pytest~=7.1.3
websocket-client~=1.4.0